% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSignalMap.R
\name{create_master_interaction_list}
\alias{create_master_interaction_list}
\title{Create Master Interaction List}
\usage{
create_master_interaction_list(
  enrichr_results = results$enrichr_results,
  de_receptors = results$upreg_receptors_filtered_and_compared,
  scSignalMap_data_filtered = results$interactions_filtered
)
}
\arguments{
\item{enrichr_results}{Data frame of Enrichr pathway enrichment results, default is results$enrichr_results}

\item{de_receptors}{Data frame of upregulated receptors, default is results$upreg_receptors_filtered_and_compared}

\item{scSignalMap_data_filtered}{Data frame of filtered interactions, default is results$interactions_filtered}
}
\value{
A data frame containing merged ligand/receptor info, enrichment, and interaction data
}
\description{
This function creates master interaction list by combining DE ligands/receptors, Enrichr results, and scSignalMap interactions
}
