% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSignalMap.R
\name{filter_lr_interactions}
\alias{filter_lr_interactions}
\title{Filter identified ligand-receptor interactions}
\usage{
filter_lr_interactions(
  interactions = NULL,
  sender_celltypes = NULL,
  receiver_celltypes = NULL,
  secreted_lig = TRUE
)
}
\arguments{
\item{interactions:}{ligand-receptor interactions dataframe from MapInteractions function of scSignalMap}

\item{sender_celltypes:}{celltype of cell(s) that are "senders" in the cell-to-cell communication interaction, ex. "\if{html}{\out{<celltype>}}" or c("celltype1", "celltype2")}

\item{receiver_celltypes:}{celltype of cell(s) that are "receivers" in the cell-to-cell communication interaction, ex. "\if{html}{\out{<celltype>}}" or c("celltype1", "celltype2")}

\item{secreted_lig:}{logical statement if the ligand in the interaction is secreted by the sending cell, default is TRUE}
}
\value{
A data frame of filtered interactions
}
\description{
This function filters interactions from a scSignalMap output based on celltype and ligand secretion
}
