% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSignalMap.R
\name{find_enriched_pathways}
\alias{find_enriched_pathways}
\title{Identify Enriched Pathways from DE Genes Using Enrichr}
\usage{
find_enriched_pathways(
  seurat_obj = NULL,
  de_condition_filtered = NULL,
  enrichr_databases = c("BioCarta_2016", "GO_Biological_Process_2025", "KEGG_2021_Human",
    "NCI-Nature_2016", "WikiPathways_2024_Human"),
  adj_p_val_method = "BH",
  adj_p_val_cutoff = 0.05
)
}
\arguments{
\item{seurat_obj:}{seurat object containing genes for Enrichr to use as background, likely to be seurat object used as input for scSignalMap's MapInteractions function, must provide input}

\item{de_condition_filtered:}{DE genes output from find_markers_btwn_cond_for_celltype function, used to extract genes for Enrichr use}

\item{enrichr_databases:}{databases that Enrichr will use for pathway idenification, default includes c("BioCarta_2016", "GO_Biological_Process_2025", "KEGG_2021_Human", "NCI-Nature_2016", "WikiPathways_2024_Human")}

\item{adj_p_val_method:}{method for p-value adjustments, default is Benjamini-Hochberg (BH)}

\item{adj_p_val_cutoff:}{desired cutoff for the adjusted p-value, default is 0.05}
}
\value{
A data frame containing identified pathways, associated statistical values, common genes, etc.
}
\description{
This function performs pathway enrichment analysis on differentially expressed (DE) genes
using the Enrichr. It filters enrichment results by an adjusted p-value cutoff and
saves both individual and combined results for selected databases. The function uses a
user-provided background gene set (from a Seurat object) and compares enrichment results
to input DE genes to highlight relevant pathways.
}
