% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSignalMap.R
\name{find_markers_btwn_cond_for_celltype}
\alias{find_markers_btwn_cond_for_celltype}
\title{Perform Differential Expression for a Desired Celltype}
\usage{
find_markers_btwn_cond_for_celltype(
  seurat_obj = NULL,
  prep_SCT = FALSE,
  cond_column = NULL,
  cond_name1 = NULL,
  cond_name2 = NULL,
  celltype_column = NULL,
  celltype_name = NULL,
  FC_cutoff = 0.3,
  adj_p_val_cutoff = 0.05
)
}
\arguments{
\item{seurat_obj:}{seurat object containing scRNA data}

\item{prep_SCT:}{specify whether the PrepSCTFindMarker function needs to be applied, defaults to FALSE}

\item{cond_column:}{meta.data column name for condition information}

\item{cond_name1:}{condition column value that compared to cond_name2 for differential expression analysis, must be specified}

\item{cond_name2:}{condition column value that compared to cond_name1 for differential expression analysis, must be specified}

\item{celltype_column:}{meta.data column name for celltype information}

\item{celltype_name:}{celltype column value of interest, must be specified}

\item{FC_cutoff:}{desired cutoff for log2FC values using >= the absolute value, default is 0.3}

\item{adj_p_val_cutoff:}{desired cutoff for the adjusted p value, default is 0.05}
}
\value{
A data frame of DE genes between conditions for specified cell type, including Ensembl ID, gene symbol, log2FC, and adjusted p-value.
}
\description{
This function identifies differentially expressed (DE) genes between two conditions
within a specific celltype of interest. It uses Seurat's \code{FindMarkers} function with
customizable log2FC and adjusted p-value cutoffs. The results are filtered and
annotated with gene symbols from Ensembl IDs.
}
