% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSignalMap.R
\name{find_upreg_receptors}
\alias{find_upreg_receptors}
\title{Identify upregulated receptors}
\usage{
find_upreg_receptors(de_condition_filtered = NULL, FC_cutoff = 0.3)
}
\arguments{
\item{de_condition_filtered:}{differentially expressed genes output from find_markers_btwn_cond_for_celltype function}

\item{FC_cutoff:}{desired cutoff for log2FC values using >= the absolute value, default is 0.3}
}
\value{
A dataframe with identified DE genes and their log2FC from previously chosen condition
}
\description{
This function identifies upregualted receptors from a given DE gene table using a chosen log2FC cutoff
}
