% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSignalMap.R
\name{intersect_upreg_receptors_with_lr_interactions}
\alias{intersect_upreg_receptors_with_lr_interactions}
\title{Intersect upregulated receptors with filtered ligand-receptor interactions}
\usage{
intersect_upreg_receptors_with_lr_interactions(
  upreg_receptors = NULL,
  interactions = NULL
)
}
\arguments{
\item{upreg_receptors:}{name for output file containing DE receptors}

\item{interactions:}{ligand-receptor interactions dataframe from map_interactions}
}
\value{
A data frame of idendified DE receptor genes found in previously idendified interactions list
}
\description{
This function takes previously identified DE receptor genes from chosen condition and identifies overlapping interactions
}
