% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSignalMap.R
\name{MapInteractions}
\alias{MapInteractions}
\title{Capturing signaling pathways using scRNA-seq data}
\usage{
MapInteractions(
  seurat_obj,
  group_by,
  avg_log2FC_gte = 0.25,
  p_val_adj_lte = 0.05,
  min_pct = 0.1,
  species = "human",
  gene_id = "ensembl"
)
}
\arguments{
\item{seurat_obj:}{a seurat object must be supplied to classify, no default}

\item{group_by:}{a meta_data column that will be used to split the cells, no default}

\item{avg_log2FC_gte:}{}

\item{p_val_adj_lte:}{}

\item{species:}{from which species did the samples originate, either 'human' or 'mouse', defaults to 'human'}

\item{gene_id:}{what type of gene ID is used, either 'ensembl' or 'symbol', defaults to 'ensembl'}
}
\value{
data.frame with putative interactions
}
\description{
Function to capture signaling pathways using scRNA-seq data.
}
