% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSignalMap.R
\name{run_full_scSignalMap_pipeline}
\alias{run_full_scSignalMap_pipeline}
\title{Run Full scSignalMap Pipeline}
\usage{
run_full_scSignalMap_pipeline(
  workingdir = "/files",
  seurat_obj = NULL,
  prep_SCT = TRUE,
  cond_column = NULL,
  cond_name1 = NULL,
  cond_name2 = NULL,
  celltype_column = NULL,
  celltype_name = NULL,
  sender_celltypes = NULL,
  receiver_celltypes = NULL,
  secreted_lig = TRUE,
  FC_cutoff = 0.3,
  adj_p_val_cutoff = 0.05,
  enrichr_databases = c("BioCarta_2016", "GO_Biological_Process_2025", "KEGG_2021_Human",
    "NCI-Nature_2016", "WikiPathways_2024_Human"),
  adj_p_val_method = "BH"
)
}
\arguments{
\item{workingdir:}{working directory where results will be saved, default is "/files"}

\item{seurat_obj:}{seurat object containing scRNA-seq data}

\item{prep_SCT:}{logical; whether to run PrepSCTFindMarkers before differential expression, default = TRUE}

\item{cond_column:}{meta.data column in Seurat object containing condition labels}

\item{cond_name1:}{condition column value that compared to cond_name2 for differential expression analysis, must be specified}

\item{cond_name2:}{condition column value that compared to cond_name1 for differential expression analysis, must be specified}

\item{celltype_column:}{meta.data column in Seurat object for celltype information}

\item{celltype_name:}{celltype column value of interest, must be specified}

\item{sender_celltypes:}{celltype of cell(s) that are "senders" in the cell-to-cell communication interaction, ex. "\if{html}{\out{<celltype>}}" or c("celltype1", "celltype2")}

\item{receiver_celltypes:}{celltype of cell(s) that are "receivers" in the cell-to-cell communication interaction, ex. "\if{html}{\out{<celltype>}}" or c("celltype1", "celltype2")}

\item{secreted_lig:}{logical; whether to restrict to secreted ligands in filtering, default = TRUE}

\item{FC_cutoff:}{log2 fold change cutoff for differential expression and receptor filtering, default = 0.3}

\item{adj_p_val_cutoff:}{adjusted p-value cutoff for differential expression and pathway enrichment, default = 0.05}

\item{enrichr_databases:}{databases that Enrichr will use for pathway idenification, default includes c("BioCarta_2016", "GO_Biological_Process_2025", "KEGG_2021_Human", "NCI-Nature_2016", "WikiPathways_2024_Human")}

\item{adj_p_val_method:}{method for p-value adjustments, default = "BH"}
}
\value{
a list containing ligand–receptor interactions, DE genes, upregulated receptors, filtered interactions, intersected receptors, and pathway enrichment results.
}
\description{
This function performs a full ligand–receptor signaling analysis
using scSignalMap and Seurat. It performs ligand–receptor interaction mapping,
differential expression analysis, receptor filtering, receptor–ligand
intersection, and pathway enrichment analysis in one workflow.
}
